refactor(observability): consolidate metrics types in metrics.types.ts
Move types that were co-located with classes/functions into a dedicated
src/metrics/metrics.types.ts (mirrors how src/otel/config.ts holds OTel
types). Affected:

- PromClientRuntime, JetstreamMetrics, CreateMetricsOptions (was metrics.factory.ts)
- ConsumerPollTarget, PollRunnerOptions, PollErrorTarget (was poll-runner.ts)
- ResolvedSubjectLabels, RecoveredKindLabel (was metrics.service.ts)

Implementation files keep imports; tests updated. No behavior change.
