+++ src/metrics/metrics.factory.ts
@@ -1,51 +1,7 @@
-import type { Counter, Histogram, Gauge, Registry } from 'prom-client';
+import type { Counter, Gauge, Histogram } from 'prom-client';
 
 import { DEFAULT_HISTOGRAM_BUCKETS, DEFAULT_METRICS_PREFIX } from './metrics.constants';
-import type { HistogramBuckets } from './metrics.config';
-
-/**
- * Subset of the `prom-client` runtime surface used by the factory. Injected
- * (rather than statically imported) so nothing here triggers a load until
- * `JetstreamMetricsModule` resolves the peer dependency via dynamic import.
- */
-/* eslint-disable @typescript-eslint/naming-convention */
-export interface PromClientRuntime {
-  Counter: typeof Counter;
-  Histogram: typeof Histogram;
-  Gauge: typeof Gauge;
-}
-/* eslint-enable @typescript-eslint/naming-convention */
-
-/** Resolved metric handles. Names are stable; emitters use them by property. */
-export interface JetstreamMetrics {
-  messagesReceivedTotal: Counter<string>;
-  messagesProcessedTotal: Counter<string>;
-  messagesUnhandledTotal: Counter<string>;
-  messagesDeadLetterTotal: Counter<string>;
-  publishTotal: Counter<string>;
-  rpcTimeoutTotal: Counter<string>;
-  consumerRecoveredTotal: Counter<string>;
-  errorsTotal: Counter<string>;
-  handlerDurationSeconds: Histogram<string>;
-  publishDurationSeconds: Histogram<string>;
-  rpcDurationSeconds: Histogram<string>;
-  consumerNumPending: Gauge<string>;
-  consumerNumAckPending: Gauge<string>;
-  consumerNumRedelivered: Gauge<string>;
-  consumerNumWaiting: Gauge<string>;
-  streamMessages: Gauge<string>;
-  streamBytes: Gauge<string>;
-  connectionUp: Gauge<string>;
-  metricsPollErrorsTotal: Counter<string>;
-}
-
-export interface CreateMetricsOptions {
-  register: Registry;
-  promClient: PromClientRuntime;
-  prefix?: string;
-  defaultLabels?: Record<string, string>;
-  buckets?: HistogramBuckets;
-}
+import type { CreateMetricsOptions, JetstreamMetrics } from './metrics.types';
 
 /** Instantiate every Jetstream metric on the given registry. */
 export const createMetrics = (opts: CreateMetricsOptions): JetstreamMetrics => {
